requirements.txt
PyPDF2~=3.0.1
pytesseract~=0.3.13
pillow~=11.1.0
tika~=2.6.0
requests~=2.32.3